fixes to accomodate strong type checking by MatlabHandler
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@5748 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/matlab/OME/Statistics/ChebyshevStatistics.m
@@ -48,5 +48,5 @@ function ch = ChebyshevStatistics(Im,N);
 if nargin < 2 N = 20; end
 
 ch = ChebyshevTransform(Im,N);
-ch = hist(ch(:),32);
+ch = uint16(hist(ch(:),32));
 return;
